+++ src/app/map/[monitor_id]/datasets/page.tsx
@@ -1,30 +1,30 @@
-'use client';
+import axios from 'axios';
+import type { Metadata, NextPage } from 'next';
 
-import type { NextPage } from 'next';
+import type { Monitor } from '@/types/monitors';
 
-import { useMonitorLayers } from '@/hooks/monitors';
+import DatasetPageComponent from '@/components/datasets/page';
 
-import DatasetCard from '@/components/datasets/card';
-import Loading from '@/components/loading';
+type Props = {
+  params: { monitor_id: string };
+};
+
+export async function generateMetadata({ params }: Props): Promise<Metadata> {
+  // read route params
+  const id = params.monitor_id;
+
+  // fetch data
+  const monitorData = await axios
+    .get<Monitor[]>(`${process.env.NEXT_PUBLIC_API_URL}monitors/${id}`)
+    .then((response) => response.data);
+
+  return {
+    title: `${monitorData[0].title} - Open Earth Monitor Cyberinfrastructure`,
+  };
+}
 
 const DatasetsPage: NextPage<{ params: { monitor_id: string } }> = ({ params: { monitor_id } }) => {
-  const { data, isLoading, isFetched, isError } = useMonitorLayers({ monitor_id });
-  return (
-    <div>
-      {isLoading && <Loading />}
-      {isFetched && !isError && (
-        <ul className="space-y-6" data-testid="datasets-list">
-          {data.map((dataset, index) => {
-            return (
-              <li key={dataset.layer_id}>
-                <DatasetCard {...dataset} id={dataset.layer_id} defaultActive={index === 0} />
-              </li>
-            );
-          })}
-        </ul>
-      )}
-    </div>
-  );
+  return <DatasetPageComponent monitor_id={monitor_id} />;
 };
 
 export default DatasetsPage;