+++ src/app/map/page.tsx
@@ -9,7 +9,6 @@ const StaticMap = dynamic(() => import('@/components/map/static'), { ssr: false 
 
 export const metadata: Metadata = {
   title: 'Map - Open Earth Monitor Cyberinfrastructure',
-  description: '...',
 };
 
 const MapLayout: NextPage = () => (