+++ src/app/page.tsx
@@ -7,7 +7,8 @@ import PreFooter from '@/components/pre-footer';
 
 export const metadata: Metadata = {
   title: 'Hub - Open Earth Monitor Cyberinfrastructure',
-  description: '...',
+  description:
+    'Open Earth Monitor Cyberinfrastructure is an ecosystem of actors creating and using data tools in support of the sustainable environmental policy.',
 };
 
 const Hub: NextPage = () => (